requirements.txt
# Основные зависимости Django
asgiref==3.9.2
Django==5.2.6
djangorestframework==3.16.1
psycopg2-binary==2.9.10
python-decouple==3.8
sqlparse==0.5.3
tzdata==2025.2

# Тестирование
pytest==7.4.3
pytest-django==4.7.0
pytest-cov==4.1.0
pytest-xdist==3.3.1

# Дополнительные инструменты для тестирования
factory-boy==3.3.0
freezegun==1.2.2
